# openapi/api.yaml
openapi: 3.1.0
info:
  title: Prompt Vault API
  version: 0.1.0
  description: Shared contract for the Prompt Vault auth slice.
servers:
  - url: /
paths:
...
  /api/prompts:
    post:
      tags:
        - Prompts
      operationId: createPrompt
      summary: Create a private prompt
      description: Creates a private prompt owned by the current authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromptRequest'
      responses:
        '201':
          description: Prompt created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: No authenticated session is present.
  /api/prompts/{promptId}/submissions:
    post:
      tags:
        - Prompts
      operationId: SubmitPromptRequest
      summary: Create a submission for this prompt
      description: Creates an entry in the prompt submission history
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPromptRequest'
      responses:
        '201':
          description: Submission history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitPromptResponse'
        '401':
          description: No authenticated session is present.
        '403':
          description: Not authorized to submit this prompt.
    get:
      tags:
        - Prompts
      operationId: listPromptsSubmission
      summary: Get all the submission responses for a prompt
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Submission history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmitPromptResponse'
        '400':
          description: Validation failed
        '401':
          description: No authenticated session is present.
        '403':
          description: Not authorized to retrieve submissions.
  /api/public-prompts:
    get:
      tags:
        - Prompts
      operationId: listPublicPrompts
      summary: List public prompts
      description: Returns public prompts owned by other enabled users.
      responses:
        '200':
          description: Public prompts visible to the current authenticated user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicPrompt'
        '401':
          description: No authenticated session is present.
  /api/public-prompts/{promptId}:
    get:
      tags:
        - Prompts
      operationId: getPublicPrompt
      summary: Get public prompt detail
      description: Returns public prompt detail when the prompt is owned by another enabled user.
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Public prompt detail visible to the current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPrompt'
        '401':
          description: No authenticated session is present.
        '404':
          description: The prompt is not visible as a public prompt to the current user.
  /api/prompts/{promptId}:
    get:
      tags:
        - Prompts
      operationId: getPrompt
      summary: Get an owned prompt
      description: Returns prompt detail when the prompt is owned by the current authenticated user.
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Prompt detail for the current owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
        '401':
          description: No authenticated session is present.
        '404':
          description: The prompt does not exist for the current owner.
    patch:
      tags:
        - Prompts
      operationId: updatePrompt
      summary: Update an owned prompt
      description: Updates prompt content and category when the prompt is owned by the current authenticated user.
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromptRequest'
      responses:
        '200':
          description: The updated prompt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: No authenticated session is present.
        '404':
          description: The prompt does not exist for the current owner.
    delete:
      tags:
        - Prompts
      operationId: deletePrompt
      summary: Delete an owned prompt
      description: Permanently deletes a prompt when it is owned by the current authenticated user.
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Prompt deleted successfully.
        '401':
          description: No authenticated session is present.
        '404':
          description: The prompt does not exist for the current owner.
  /api/prompts/{promptId}/visibility:
    patch:
      tags:
        - Prompts
      operationId: updatePromptVisibility
      summary: Share or unshare an owned prompt
      description: Updates Prompt Visibility when the prompt is owned by the current authenticated user.
      parameters:
        - name: promptId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePromptVisibilityRequest'
      responses:
        '200':
          description: The prompt with the requested visibility.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prompt'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: No authenticated session is present.
        '404':
          description: The prompt does not exist for the current owner.
  /api/users/{userId}/prompts:
    get:
      tags:
        - Prompts
      operationId: listMyPrompts
      summary: List my prompts
      description: Returns all prompts owned by the current authenticated user, including private prompts.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Prompts owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prompt'
        '401':
          description: No authenticated session is present.
        '403':
          description: The requested user is not the current authenticated user.
...
  /api/admin/policy/keywords:
    get:
      tags:
        - AdminPolicyKeywords
      operationId: listPolicyKeywords
      summary: List Policy Keywords
      description: Returns the Policy Keyword catalog for administrators.
      responses:
        '200':
          description: Policy Keywords available for administration.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PolicyKeyword'
        '401':
          description: No authenticated session is present.
        '403':
          description: The authenticated user is not an administrator.
    post:
      tags:
        - AdminPolicyKeywords
      operationId: createPolicyKeyword
      summary: Create a Policy Keyword
      description: Creates a globally managed Policy Keyword. Admin users only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyKeywordRequest'
      responses:
        '201':
          description: Policy Keyword created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyKeyword'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: No authenticated session is present.
        '403':
          description: The authenticated user is not an administrator.
  /api/admin/policy/keywords/{keywordId}:
    patch:
      tags:
        - AdminPolicyKeywords
      operationId: updatePolicyKeyword
      summary: Update a Policy Keyword
      description: Updates an existing globally managed Policy Keyword. Admin users only.
      parameters:
        - name: keywordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyKeywordRequest'
      responses:
        '200':
          description: Policy Keyword updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyKeyword'
        '400':
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: No authenticated session is present.
        '403':
          description: The authenticated user is not an administrator.
        '404':
          description: Policy Keyword was not found.
    delete:
      tags:
        - AdminPolicyKeywords
      operationId: deletePolicyKeyword
      summary: Delete a Policy Keyword
      description: Deletes an existing globally managed Policy Keyword. Admin users only.
      parameters:
        - name: keywordId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Policy Keyword deleted successfully.
        '401':
          description: No authenticated session is present.
        '403':
          description: The authenticated user is not an administrator.
        '404':
          description: Policy Keyword was not found.
components:
  schemas:
...
    PolicyKeyword:
      type: object
      additionalProperties: false
      required:
        - id
        - keyword
        - createdAt
        - createdByUserId
        - createdByUsername
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        keyword:
          type: string
        createdAt:
          type: string
          format: date-time
        createdByUserId:
          type: integer
          format: int64
        createdByUsername:
          type: string
          description: Username of the Admin who created the Policy Keyword.
        updatedAt:
          type: string
          format: date-time
    PolicyKeywordRequest:
      type: object
      additionalProperties: false
      required:
        - keyword
      properties:
        keyword:
          type: string
          minLength: 1
          maxLength: 100
          description: Policy Keyword text is edge-trimmed before validation and persistence while preserving internal whitespace and punctuation.
...
    PromptVisibility:
      type: string
      enum:
        - PRIVATE
        - PUBLIC
    CreatePromptRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - text
        - categoryId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          description: Prompt titles are trimmed before validation and persistence.
        text:
          type: string
          minLength: 1
          maxLength: 10000
          description: Prompt text is edge-trimmed before validation and persistence while preserving internal whitespace and newlines.
        categoryId:
          type: integer
          format: int64
          minimum: 1
        visibility:
          type: string
          default: PRIVATE
          enum:
            - PUBLIC
            - PRIVATE
    UpdatePromptRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - text
        - categoryId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          description: Prompt titles are trimmed before validation and persistence.
        text:
          type: string
          minLength: 1
          maxLength: 10000
          description: Prompt text is edge-trimmed before validation and persistence while preserving internal whitespace and newlines.
        categoryId:
          type: integer
          format: int64
          minimum: 1
    UpdatePromptVisibilityRequest:
      type: object
      additionalProperties: false
      required:
        - visibility
      properties:
        visibility:
          $ref: '#/components/schemas/PromptVisibility'
    Prompt:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - text
        - visibility
        - categoryId
        - ownerUserId
        - createdAt
        - updatedAt
        - flaggedAt
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        text:
          type: string
        visibility:
          $ref: '#/components/schemas/PromptVisibility'
        categoryId:
          type: integer
          format: int64
        ownerUserId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        flaggedAt:
          type: [ 'null', string ]
          format: date-time
          description: Present when the Prompt was flagged by Policy Keyword matching. Matched keyword snapshots are not exposed in owner-facing responses.
    PublicPrompt:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - text
        - visibility
        - categoryId
        - ownerUsername
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        text:
          type: string
        visibility:
          $ref: '#/components/schemas/PromptVisibility'
        categoryId:
          type: integer
          format: int64
        ownerUsername:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubmitPromptRequest:
      type: object
      additionalProperties: false
      required:
        - response
      properties:
        response:
          type: string
    SubmitPromptResponse:
      type: object
      additionalProperties: false
      required:
        - response
        - submittedAt
      properties:
        response:
          type: string
        submittedAt:
          type: string
          format: date-time
...
    ValidationFieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - message
        - fieldErrors
      properties:
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationFieldError'
...